BigDecimal completamente implementado

+++ src/main/java/ufjf/poo/model/pedido/ItemPedido.java
@@ -1,6 +1,7 @@
 package ufjf.poo.model.pedido;
 
+import java.math.BigDecimal;
 import ufjf.poo.model.estoque.Produto;
 
-public record ItemPedido(Produto produto, int quantidade, long subtotal) {
+public record ItemPedido(Produto produto, int quantidade, BigDecimal subtotal) {
 }
